+++ src/app/tf/[species]/[factor]/deeplearnedselexmotif/motifenrichmentselex.tsx
@@ -26,19 +26,12 @@ import {
   Checkbox,
   Divider,
   SelectChangeEvent,
-  useMediaQuery,
   useTheme,
   Stack,
-  Chip,
 } from "@mui/material";
 import Grid from "@mui/material/Unstable_Grid2"; // Grid2 for MUI v2
 import { ApiContext } from "@/apicontext";
 import { Logo, DNAAlphabet } from "logojs-react";
-import { scaleLinear, scaleBand } from "@visx/scale";
-import { Group } from "@visx/group";
-import { LinePath, Bar } from "@visx/shape";
-import { AxisBottom, AxisLeft } from "@visx/axis";
-import { Text } from "@visx/text";
 import SaveAltIcon from "@mui/icons-material/SaveAlt";
 import SwapHorizIcon from "@mui/icons-material/SwapHoriz";
 import {
@@ -52,20 +45,12 @@ import {
 import { downloadData, downloadSVGElementAsSVG } from "@/utilities/svgdata";
 import { meme, MMotif } from "@/components/motifsearch/motifutil";
 import { reverseComplement as rc } from "@/components/tf/geneexpression/utils";
+import SelexLinePlot from "./lineplot";
+import SelexBarPlot from "./barplot";
 
 // Add custom colors to Alphabet A and T
 DNAAlphabet[0].color = "#228b22";
 DNAAlphabet[3].color = "red";
-
-const colors: { [key: number]: string } = {
-  1: "#FFA500",
-  2: "#FF0000",
-  3: "#008000",
-  4: "#0000FF",
-  5: "#A52A2A",
-  6: "#FFD700",
-  7: "#90EE90",
-};
 
 const DeepLearnedSelexMotifs: React.FC<{ factor: string; species: string }> = ({
   factor,
@@ -367,95 +352,7 @@ const DeepLearnedSelexMotif: React.FC<{
     selex?: any;
   }[];
 }> = ({ study, assay, protein_type, data }) => {
-  const points = useMemo(
-    () =>
-      data.flatMap((d) =>
-        d.roc_curve.map((r) => ({
-          x: r[0],
-          y: r[1],
-          round: d.selex_round,
-        }))
-      ),
-    [data]
-  );
-
-  const domain = useMemo(
-    () => ({
-      x: {
-        start: 0,
-        end: 1,
-      },
-      y: {
-        start: 0,
-        end: 1,
-      },
-    }),
-    []
-  );
-
-  const barplotDomain = useMemo(
-    () => ({
-      x: {
-        start: Math.min(...data.map((x) => x.selex_round)) - 1,
-        end: Math.max(...data.map((x) => x.selex_round)) + 1,
-      },
-      y: {
-        start: 0.0,
-        end: Math.max(...data.map((x) => x.fractional_enrichment)) * 1 + 0.1,
-      },
-    }),
-    [data]
-  );
-
-  const lineGraphHeight = 300;
-  const lineGraphWidth = 300;
-  const barGraphHeight = 300;
-  const barGraphWidth = 300;
-  const margin = { top: 20, right: 90, bottom: 70, left: 70 };
-
   const theme = useTheme();
-  const isMobile = useMediaQuery(theme.breakpoints.down("sm"));
-
-  const xScale = useMemo(
-    () =>
-      scaleLinear({
-        domain: [domain.x.start, domain.x.end],
-        range: [3, lineGraphWidth - (isMobile ? 0 : margin.right)],
-      }),
-    [domain, lineGraphWidth, margin, isMobile]
-  );
-
-  const yScale = useMemo(
-    () =>
-      scaleLinear({
-        domain: [domain.y.start, domain.y.end],
-        range: [lineGraphHeight - margin.bottom, margin.top],
-      }),
-    [domain, lineGraphHeight, margin]
-  );
-
-  const barXScale = useMemo(
-    () =>
-      scaleBand({
-        domain: data.map((d) => d.selex_round),
-        range: [0, barGraphWidth - (isMobile ? 0 : margin.right)],
-        paddingInner: 0.5,
-        paddingOuter: 0.3,
-      }),
-    [data, barGraphWidth, margin, isMobile]
-  );
-
-  const barYScale = useMemo(
-    () =>
-      scaleLinear({
-        domain: [barplotDomain.y.start, barplotDomain.y.end],
-        range: [barGraphHeight - margin.bottom, margin.top],
-      }),
-    [barplotDomain, barGraphHeight, margin]
-  );
-
-  const lineref = useRef<SVGSVGElement | null>(null);
-  const barref = useRef<SVGSVGElement | null>(null);
 
   const downloadSVGElement = (
     ref: MutableRefObject<SVGSVGElement | null>,
@@ -466,12 +363,8 @@ const DeepLearnedSelexMotif: React.FC<{
     }
   };
 
-
-  const presentCycles = data.map((d) => d.selex_round);
-
-
   return (
-    <Box sx={{ padding: "1em" }}>
+    <Box>
       <Typography variant="h5" align="left" mt={2} gutterBottom>
         <b>
           {assay.replaceAll("-", " ")} motifs for{" "}
@@ -481,265 +374,43 @@ const DeepLearnedSelexMotif: React.FC<{
           found in {study.replace("_", " ")} study
         </b>
       </Typography>
-      <Grid container spacing={3}>
-        <Grid >
+      <Stack
+        divider={<Divider />}
+      >
+        <Grid container spacing={3}>
+          {/* LINE PLOT */}
+          <Grid xs={3}>
+            <SelexLinePlot
+              data={data}
+              downloadSVGElement={downloadSVGElement}
+            />
+          </Grid>
+          {/* BAR PLOT */}
+          <Grid xs={3}>
+            <SelexBarPlot
+              data={data}
+              downloadSVGElement={downloadSVGElement}
+            />
+          </Grid>
+        </Grid>
+        {/* CYCLES */}
+        <Grid container spacing={3}>
           {data.map((d, i) => (
-            <Box key={`logo${i}`} sx={{ textAlign: "center" }}>
-              <Typography variant="h6" sx={{ color: "brown" }}>
-                Cycle {d.selex_round}
-              </Typography>
-              {d.ppm && d.ppm.length > 0 && (
-                <>
-                  <DownloadableMotif ppm={d.ppm} name={study} />
-
-                </>
-              )}
-            </Box>
+            <Grid xs={3}>
+              <Box key={`logo${i}`} sx={{ textAlign: "center" }}>
+                <Typography variant="h6" sx={{ color: "brown" }}>
+                  Cycle {d.selex_round}
+                </Typography>
+                {d.ppm && d.ppm.length > 0 && (
+                  <>
+                    <DownloadableMotif ppm={d.ppm} name={study} />
+                  </>
+                )}
+              </Box>
+            </Grid>
           ))}
         </Grid>
-        <Grid md={5}>
-          <Box
-            sx={{
-              display: "flex",
-              flexDirection: "column",
-              alignItems: isMobile ? "flex-start" : "center",
-              marginTop: "30px",
-              marginBottom: "20px"
-            }}
-          >
-            <Button
-              variant="contained"
-              startIcon={<SaveAltIcon />}
-              onClick={() => downloadSVGElement(lineref, "lineplot.svg")}
-              sx={{
-                marginTop: "20px",
-              }}
-            >
-              Download Line Plot
-            </Button>
-            <svg ref={lineref} width={lineGraphWidth} height={lineGraphHeight}>
-
-              <Group left={isMobile ? 0 : margin.left} top={margin.top}>
-
-                <AxisLeft
-                  scale={yScale}
-                  label="1-Specificity"
-                  labelProps={{
-                    fontSize: 12,
-                    fill: "black",
-                    textAnchor: "middle",
-                    transform: "translate(-60, 0) rotate(-90)",
-                  }}
-                  strokeWidth={3}
-                  tickValues={[0, 0.5, 1.0]}
-                  tickLabelProps={() => ({
-                    fontSize: 10,
-                    fill: "black",
-                    textAnchor: "end",
-                    dx: "-0.25em",
-                    dy: "0.25em",
-                  })}
-                />
-                <AxisBottom
-                  scale={xScale}
-                  top={lineGraphHeight - margin.bottom}
-                  label="Sensitivity"
-                  labelProps={{
-                    fontSize: 12,
-                    fill: "black",
-                    textAnchor: "middle",
-                    transform: "translate(0, 40)",
-                  }}
-                  strokeWidth={3}
-                  tickValues={[0, 0.5, 1.0]}
-                  tickLabelProps={() => ({
-                    fontSize: 10,
-                    fill: "black",
-                    textAnchor: "middle",
-                    dy: "0.25em",
-                  })}
-                />
-                {data.map((d, i) => (
-                  <LinePath
-                    key={i}
-                    data={points.filter((p) => p.round === d.selex_round)}
-                    x={(p) => xScale(p.x)}
-                    y={(p) => yScale(p.y)}
-                    stroke={colors[d.selex_round]}
-                    strokeWidth={2}
-                  />
-                ))}
-                <Text
-                  x={105}
-                  y={270}
-                  fontSize={14}
-                  textAnchor="middle"
-                  fill="black"
-                >
-                  Sensitivity
-                </Text>
-                <Text
-                  x={-125}
-                  y={-40}
-                  fontSize={14}
-                  textAnchor="middle"
-                  fill="black"
-                  transform="rotate(-90)"
-                >
-                  1-Specificity
-                </Text>
-              </Group>
-            </svg>
-            <Box display="flex" alignItems="center" style={{ marginBottom: "17px" }}>
-              <Typography variant="caption" component="div" style={{ display: 'flex', justifyContent: 'center', alignItems: 'center', marginLeft: '40px', fontSize: "14px" }}>
-                Cycle  {presentCycles.map((cycle) => (
-                  <div style={{
-                    marginLeft: "5px", display: 'flex',
-                    alignItems: 'center',
-                    justifyContent: 'left'
-                  }}
-                  ><div
-                    style={{
-                      width: '10px',
-                      height: '3px',
-                      backgroundColor: colors[cycle],
-                      display: 'inline-block',
-                      alignItems: 'center',
-                      justifyContent: 'center'
-                    }}
-                  > </div>
-                    <Typography style={{ marginLeft: '2px' }}
-                      fontSize={14}
-                    > {cycle} </Typography>
-                  </div>
-                ))}
-              </Typography>
-
-            </Box>
-            <Button
-              variant="contained"
-              startIcon={<SaveAltIcon />}
-              onClick={() => downloadSVGElement(barref, "barplot.svg")}
-              sx={{
-                marginTop: "20px",
-              }}
-            >
-              Download Bar Plot
-            </Button>
-
-            <svg ref={barref} width={barGraphWidth} height={barGraphHeight}>
-              <Group left={isMobile ? 0 : margin.left} top={margin.top}>
-                <AxisLeft
-                  scale={barYScale}
-                  label="Fractional Enrichment"
-                  labelProps={{
-                    fontSize: 12,
-                    fill: "black",
-                    textAnchor: "middle",
-                    transform: "translate(-60, 0) rotate(-90)",
-                  }}
-                  strokeWidth={3}
-                  tickValues={[0, data[data.length - 1].fractional_enrichment * 0.5, data[data.length - 1].fractional_enrichment, data[data.length - 1].fractional_enrichment * 1.5]}
-                  tickLabelProps={() => ({
-                    fontSize: 10,
-                    fill: "black",
-                    textAnchor: "end",
-                    dx: "-0.25em",
-                    dy: "0.25em",
-                  })}
-                />
-                <AxisBottom
-                  scale={barXScale}
-                  top={barGraphHeight - margin.bottom}
-                  label="Cycle"
-                  strokeWidth={3}
-                  labelProps={{
-                    fontSize: 12,
-                    fill: "black",
-                    textAnchor: "middle",
-                    transform: "translate(0, 40)",
-                  }}
-                  tickLabelProps={() => ({
-                    fontSize: 10,
-                    fill: "black",
-                    textAnchor: "middle",
-                    dy: "0.25em",
-                  })}
-
-                />
-                {data.map((d, i) => (
-                  <React.Fragment key={i}>
-                    <Bar
-                      x={barXScale(d.selex_round)}
-                      y={barYScale(d.fractional_enrichment)}
-                      height={
-                        barGraphHeight -
-                        margin.bottom -
-                        barYScale(d.fractional_enrichment)
-                      }
-                      width={barXScale.bandwidth()}
-                      fill={colors[d.selex_round]}
-
-                    />
-                    <Text
-                      x={barXScale(d.selex_round)! + barXScale.bandwidth() / 2}
-                      y={barYScale(d.fractional_enrichment) - 5}
-                      fontSize={12}
-                      fill={colors[d.selex_round]}
-                      textAnchor="middle"
-                    >
-                      {d.fractional_enrichment.toFixed(2)}
-                    </Text>
-                  </React.Fragment>
-                ))}
-                <Text
-                  x={100}
-                  y={270}
-                  fontSize={14}
-                  textAnchor="middle"
-                  fill="black"
-                >
-                  Cycle
-                </Text>
-                <Text
-                  x={-135}
-                  y={-40}
-                  fontSize={14}
-                  textAnchor="middle"
-                  fill="black"
-                  transform="rotate(-90)"
-                >
-                  Fractional Enrichment
-                </Text>
-              </Group>
-            </svg>
-            <Typography variant="caption" component="div" style={{ display: 'flex', justifyContent: 'center', alignItems: 'center', marginLeft: '20px', fontSize: "14px" }}>
-              Cycle  {[2, 3, 4].map((cycle, index) => (
-                <div style={{
-                  marginLeft: "5px", display: 'flex',
-                  alignItems: 'center',
-                  justifyContent: 'left',
-                }}
-                ><div
-                  style={{
-                    width: '10px',
-                    height: '8px',
-                    backgroundColor: colors[cycle],
-                    display: 'inline-block',
-                    alignItems: 'center',
-                    justifyContent: 'center'
-                  }}
-                > </div>
-                  <Typography style={{ marginLeft: '2px' }}
-                    fontSize={14}
-                  > {cycle} </Typography>
-                </div>
-              ))}
-            </Typography>
-
-          </Box>
-        </Grid>
-      </Grid>
+      </Stack>
     </Box>
   );
 };
